requirements.txt
fastapi
uvicorn
groq
python-dotenv
PyPDF2
python-docx

